refactor(files): change structure files
