--- src/containers/Files/routers.js
@@ -1,12 +1,0 @@
-import EmptyMessage from '../../components/EmptyMessage'
-
-const routes = [
-    {
-        path: '/',
-        name: 'No selected',
-        component: EmptyMessage,
-        exact: true
-    }
-]
-
-export default routes
